+++ src/scene.rs
@@ -2,6 +2,6 @@ use crate::context::Context;
 
 pub trait Scene {
     fn draw(&mut self, ctx: &mut Context);
-    fn update(&mut self, delta: f32);
+    fn update(&mut self, delta: f64);
     fn event<T>(&mut self, event: &winit::event::Event<T>) -> bool;
 }
